core-image-boot-initramfs: Add image
Some systems need to load firmware at boot time that cannot wait until the
final root filesystem is running - and also cannot be retried at a later
time by the driver.

An example of these is the SDMA firmware on the i.MX6.

This recipe provides an initramfs that includes the required firmware.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>