core-image-boot-initrams: Add imx-sdma firmware to initramfs
The i.mx6 SDMA engine comes with ROM firmware that can be patched by a
firmware file.

Once the SDMA driver is loaded, it either find a patch file or runs with
its unpatched ROM firmware.

This commit make the imx-sdma driver a module and add its firmware patch
to the initramfs so that it is correcly loaded when the system is booting.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>